Move CommonMetadata to separate module
